refactor (web): extract shared TableSearchBar component
Consolidate 3 similar SearchBar components into web/src/components.
Removes duplication from route, forward, decap modules.
