+++ web/src/components/TableSearchBar/index.ts
@@ -1,0 +1,2 @@
+export { TableSearchBar } from './TableSearchBar';
+export type { TableSearchBarProps } from './TableSearchBar';
