+++ web/src/components/index.ts
@@ -10,6 +10,7 @@ export * from './ConfirmDialog';
 export * from './CreateEntityDialog';
 export * from './SortableDataTable';
 export * from './SortableTableHeader';
+export * from './TableSearchBar';
 export * from './GraphEditor';
 export * from './CounterDisplay';
 export * from './CountersContext';
