+++ web/src/pages/route/FIBTable.tsx
@@ -1,12 +1,11 @@
 import React, { useRef, useCallback, useMemo, useState } from 'react';
 import { useVirtualizer } from '@tanstack/react-virtual';
 import { Box, Text } from '@gravity-ui/uikit';
-import { EmptyState } from '../../components';
+import { EmptyState, TableSearchBar } from '../../components';
 import type { FIBEntry, FIBNexthop } from '../../api/routes';
 import type { FIBRow, FIBSortableColumn, FIBSortState } from './types';
 import { useContainerHeight } from './hooks';
 import { ROW_HEIGHT, OVERSCAN, SEARCH_BAR_HEIGHT, HEADER_HEIGHT, FOOTER_HEIGHT, FIB_TOTAL_WIDTH, fibCellStyles } from './constants';
-import { TableSearchBar } from './TableSearchBar';
 import './route.scss';
 
 export interface FIBTableProps {
@@ -209,6 +208,8 @@ export const FIBTable: React.FC<FIBTableProps> = ({ entries }) => {
                 statsText={statsText}
                 selectedText={null}
                 placeholder="Search by prefix, MAC, or device..."
+                height={SEARCH_BAR_HEIGHT}
+                inputWidth={350}
             />
 
             <Box className="route-table__wrapper">
