+++ web/src/pages/route/VirtualizedRouteTable.tsx
@@ -1,13 +1,12 @@
 import React, { useRef, useCallback, useMemo, useState } from 'react';
 import { useVirtualizer } from '@tanstack/react-virtual';
 import { Box, Text } from '@gravity-ui/uikit';
-import { EmptyState } from '../../components';
+import { EmptyState, TableSearchBar } from '../../components';
 import type { Route } from '../../api/routes';
 import type { MockRouteGenerator } from './mockData';
 import { useSortState, useContainerHeight, sortComparators } from './hooks';
 import { ROW_HEIGHT, OVERSCAN, TOTAL_WIDTH, SEARCH_BAR_HEIGHT, HEADER_HEIGHT, FOOTER_HEIGHT } from './constants';
 import { VirtualRow } from './VirtualRow';
-import { TableSearchBar } from './TableSearchBar';
 import { TableHeader } from './TableHeader';
 import './route.scss';
 
@@ -216,6 +215,9 @@ export const VirtualizedRouteTable: React.FC<VirtualizedRouteTableProps> = ({
                 selectedText={selectedText}
                 onClearSelection={handleClearSelection}
                 helperText={helperText}
+                placeholder="Search by prefix, nexthop, or peer..."
+                height={SEARCH_BAR_HEIGHT}
+                inputWidth={350}
             />
 
             {/* Table container */}
