+++ web/src/pages/route/index.ts
@@ -24,7 +24,6 @@ export * from './FIBTable';
 export * from './RouteConfigContent';
 export * from './VirtualizedRouteTable';
 export * from './VirtualRow';
-export * from './TableSearchBar';
 export * from './TableHeader';
 
 // Mock data utilities
